pyproject.toml
[project]
name = "test-is-invalid-mage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"